server response

+++ protocol/src/main/java/org/chivqsss/commands/CommandCodes.java
@@ -1,7 +1,7 @@
 package org.chivqsss.commands;
 
 public class CommandCodes {
-    public static final int GET = 0;
-    public static final int PUT = 1;
-    public static final int DELETE = 2;
+    public static final byte GET = 0;
+    public static final byte PUT = 1;
+    public static final byte DELETE = 2;
 }
